+ Added timeout handling to killer and unformat track routines.
+ Referenced killer and unformat track routines during remaster instead of 
  making them a special case.

+ Fixed usage output of nibread for -i (index hole) vs -I (interactive)



git-svn-id: https://c64preservation.com/svn/nibtools/trunk@268 048eaad1-c872-4173-8c30-523bb29f53ec

+++ nibtls_rt.c
@@ -22,7 +22,7 @@ static int
 cbm_handshaked_read(int toggle);
 
 int
-cbm_parallel_burst_write_track(int fd, unsigned char * buffer, int length)
+cbm_parallel_burst_write_track(int fd, unsigned char * buffer, size_t length)
 {
 	int i;
